* add combinatorics tests

+++ libna/na/combinatorics.h
@@ -1,11 +1,11 @@
 #ifndef _COMBINATORICS_H_
 #define _COMBINATORICS_H_
 
-int factorial(int n);
-int k_combinations(int n, int k);
-int k_combinations_rep(int n, int k);
-int k_permutations(int n, int k);
-int k_permutations_rep(int n, int k);
+long int factorial(int n);
+long int k_combinations(int n, int k);
+long int k_combinations_rep(int n, int k);
+long int k_permutations(int n, int k);
+long int k_permutations_rep(int n, int k);
 
 #endif /* _COMBINATORICS_H_ */
